Update fixing 302 issue and grammar in log message (#4)
* Corrected some grammar in a log

* Removed try-catch from interface.py

The calling program should be able to handle any exceptions that are
thrown, and if not, allow the original through.

* Added handling for current api trying to redir

The current API implimentation tries to redir the session back to the
home page if the requested api resource doesn't exist. Currently we
catch that and trigger a 'token not valid' error.

* Removed/redacted all log instances that print out the API token