Multi device (Breaking Changes) (#5)
* Corrected some grammar in a log

* Removed try-catch from interface.py

The calling program should be able to handle any exceptions that are
thrown, and if not, allow the original through.

* Added handling for current api trying to redir

The current API implimentation tries to redir the session back to the
home page if the requested api resource doesn't exist. Currently we
catch that and trigger a 'token not valid' error.

* Removed/redacted all log instances that print out the API token

* Updated library with multi-device fetching

Added the ability to fetch all devices belonging to an account using a
deivce id of "*". This allows us to dynamicly detect new devices on an
account and add them to home assistant

* Fixed some formatting in main

* Fixed existing tests

* Added new testing for multi packet requests